requirements.txt
flask-2.0.1
flask-cors
pandas
pymongo
levenshtein-0.12.0
openpyxl-3.0.7